adding bottom tab in n/w layer
