+++ app/src/main/java/com/karthik/splash/homescreen/bottomtab/di/BottomTabModule.kt
@@ -1,12 +1,10 @@
 package com.karthik.splash.homescreen.bottomtab.di
 
+import com.karthik.network.home.bottomtab.repository.BottomTabRepository
 import com.karthik.splash.homescreen.bottomtab.BottomTabTypes
-import com.karthik.splash.homescreen.bottomtab.network.BottomTabRepository
 import com.karthik.splash.homescreen.bottomtab.BottomTabViewModelFactory
-import com.karthik.splash.homescreen.bottomtab.network.BottomTabNetworkService
 import com.karthik.splash.misc.InternetHandler
 import com.karthik.splash.storage.MemoryCache
-import com.karthik.splash.storage.db.SplashDao
 import dagger.Module
 import dagger.Provides
 import retrofit2.Retrofit
@@ -25,12 +23,10 @@ class BottomTabModule {
     @Provides
     fun providesNetworkLayer(
             retrofit: Retrofit,
-            splashDao: SplashDao,
             cache: MemoryCache,
             internetHandler: InternetHandler
     ): BottomTabRepository {
-        val bottomNetworkService = retrofit.create(BottomTabNetworkService::class.java)
-        return BottomTabRepository(bottomNetworkService, splashDao, cache, internetHandler)
+        return BottomTabRepository(retrofit, cache, internetHandler)
     }
 
     @Provides
